+++ apps/backend/backend_cache.h
@@ -1,9 +1,10 @@
 /*
  *
   ***** BEGIN LICENSE BLOCK *****
- 
-  Copyright (C) 2009-2019 Olof Hagsand
-  Copyright (C) 2020-2022 Olof Hagsand and Rubicon Communications, LLC(Netgate)
+
+  Copyright (C) 2009-2016 Olof Hagsand and Benny Holmgren
+  Copyright (C) 2017-2019 Olof Hagsand
+  Copyright (C) 2020-2022 Olof Hagsand and Rubicon Communications, LLC (Netgate)
 
   This file is part of CLIXON.
 
@@ -24,24 +25,22 @@
   in which case the provisions of the GPL are applicable instead
   of those above. If you wish to allow use of your version of this file only
   under the terms of the GPL, and not to allow others to
-  use your version of this file under the terms of Apache License version 2, 
+  use your version of this file under the terms of Apache License version 2,
   indicate your decision by deleting the provisions above and replace them with
   the  notice and other provisions required by the GPL. If you do not delete
   the provisions above, a recipient may use your version of this file under
   the terms of any one of the Apache License version 2 or the GPL.
 
   ***** END LICENSE BLOCK *****
-
  */
 
-#ifndef _CLI_GENERATE_H_
-#define _CLI_GENERATE_H_
+#ifndef _BACKEND_CACHE_H_
+#define _BACKEND_CACHE_H_
 
 /*
  * Prototypes
  */
-int yang2cli_yspec(clixon_handle h, yang_stmt *yspec, char *treename);
-int yang2cli_grouping_wrap(cligen_handle ch, const char *name, cvec *cvt, void *arg, char **namep);
-int yang2cli_init(clixon_handle h);
+int backend_autocli_clear_cache(clixon_handle h);
+int from_client_clixon_cache(clixon_handle h, cxobj *xe, cbuf *cbret, void *arg, void *regarg);
 
-#endif  /* _CLI_GENERATE_H_ */
+#endif  /* _BACKEND_CACHE_H_ */