+++ CHANGELOG.md
@@ -22,6 +22,15 @@ Expected: February 2026
 ### Features
 
 * Add `-s` option to `clixon_cli` to disable output scrolling.
+* Changed cli cache to server-side instead of client-side
+  * Add cache protocol for yang-domains and clispecs
+    * This includes read and clear operations between cli client and backend
+* New `clixon-config@2025-12-01.yang` revision
+   * Added `CLICON_AUTOCLI_CACHE_DIR`
+* New `clixon-lib@2025-12-01.yang` revision
+   * Added `rpc clixon-cache`
+* New `clixon-autocli@2025-12-01.yang` revision
+   * Deprecated `clispec-cache` and `clispec-cache-dir`
 
 ### C/CLI-API changes on existing features